Even more streamling, new integer histograms
