+++ .vscode/settings.json
@@ -1,3 +1,7 @@
 {
-    "cmake.configureOnOpen": false
+    "cmake.configureOnOpen": false,
+    "files.associations": {
+        "*.py": "python",
+        "Plotter.C": "cpp"
+    }
 }
