+++ Train/EvalScripts/__init__.py
@@ -1,1 +1,2 @@
+from .eval_funcs import *
 from .eval import *
