+++ Train/TrainingScripts/__init__.py
@@ -1,1 +1,3 @@
 from .Callbacks import *
+from .prune import *
+from .train import *
