+++ Train/TrainingScripts/prune.py
@@ -12,7 +12,7 @@ import h5py
 import sklearn.metrics as metrics
 import vtx
 import TrainingScripts
-import EvalScripts.eval as eval
+import Train.EvalScripts.eval_funcs as eval_funcs
 import pandas as pd
 
 import yaml
