+++ Train/TrainingScripts/train.py
@@ -9,7 +9,7 @@ import glob
 import sklearn.metrics as metrics
 import vtx
 import TrainingScripts
-import EvalScripts.eval as eval
+import Train.EvalScripts.eval_funcs as eval_funcs
 import pandas as pd
 
 import yaml
@@ -32,10 +32,10 @@ kf = sys.argv[1]
 
 if kf == "NewKF":
     z0 = 'trk_z0'
-    bit_z0 = 'bit_trk_z0'
 elif kf == "OldKF":
     z0 = 'corrected_trk_z0'
-    bit_z0 = 'bit_corrected_trk_z0'
+elif kf == "OldKF_intZ":
+    z0 = 'corrected_int_z0'
 
 SMALL_SIZE = 15
 MEDIUM_SIZE = 15
@@ -76,7 +76,7 @@ def setup_pipeline(fileList):
     
     return ds
 
-def train_model(model,experiment,train_files,val_files,trackfeat,weightfeat,epochs=50,callbacks=None,nlatent=0,trainable=False,bit=False,model_name=None):
+def train_model(model,experiment,train_files,val_files,trackfeat,weightfeat,epochs=50,callbacks=None,nlatent=0,bit=False,model_name=[None,None]):
 
     total_steps = 0
     early_stop_patience = 100
@@ -94,93 +94,64 @@ def train_model(model,experiment,train_files,val_files,trackfeat,weightfeat,epoc
         print ("Epoch %i"%epoch)
         
         if epoch>0:
-            model.load_weights(model_name+".tf")
+            model.load_weights(model_name[0]+model_name[1]+".tf")
         
         for step,batch in enumerate(setup_pipeline(train_files)):
-            z0Shift = np.random.normal(0.0,1.0,size=batch['pvz0'].shape)
-            z0Flip = 2.*np.random.randint(2,size=batch['pvz0'].shape)-1.
-            batch[z0]=batch[z0]*z0Flip + z0Shift
-            batch['pvz0']=batch['pvz0']*z0Flip + z0Shift
+
+            if not bit:
+                z0Shift = np.random.normal(0.0,1.0,size=batch['pvz0'].shape)
+                z0Flip = 2.*np.random.randint(2,size=batch['pvz0'].shape)-1.
+                batch[z0]=batch[z0]*z0Flip + z0Shift
+                batch['pvz0']=batch['pvz0']*z0Flip + z0Shift
 
             trackFeatures = np.stack([batch[feature] for feature in trackfeat],axis=2)
 
             WeightFeatures = np.stack([batch[feature] for feature in weightfeat ],axis=2)
 
             nBatch = batch['pvz0'].shape[0]
-
-            if bit:
-                result = model.train_on_batch([batch[bit_z0],WeightFeatures,trackFeatures], [batch['pvz0'],batch['trk_fromPV'],np.zeros([nBatch,max_ntracks,1])])   
-                                              # convs                          #softmax     #binweight                                           
-            else:
-                result = model.train_on_batch([batch[z0],WeightFeatures,trackFeatures], [batch['pvz0'],batch['trk_fromPV'],np.zeros([nBatch,max_ntracks,1])])   
+                                  
+            result = model.train_on_batch([batch[z0],WeightFeatures,trackFeatures], [batch['pvz0'],batch['trk_fromPV'],np.zeros([nBatch,max_ntracks,1])])   
 
             result = dict(zip(model.metrics_names,result))
 
-            if trainable == "FH":
-                experiment.log_metric("loss",result['loss'],step=total_steps,epoch=epoch)
-                experiment.log_metric("z0_loss",result['binto_vertex_loss'],step=total_steps,epoch=epoch)
-                experiment.log_metric("assoc_loss",result['association_final_loss'],step=total_steps,epoch=epoch)
-            elif (trainable == "DiffArgMax"):
-                if nlatent > 0:
-                    experiment.log_metric("z0_loss",result['split_latent_loss'],step=total_steps,epoch=epoch)
-                else:
-                    experiment.log_metric("z0_loss",result['position_final_loss'],step=total_steps,epoch=epoch)
-                experiment.log_metric("loss",result['loss'],step=total_steps,epoch=epoch)
-                experiment.log_metric("assoc_loss",result['association_final_loss'],step=total_steps,epoch=epoch)
-
-            elif (trainable == "QDiffArgMax"):
-                if nlatent > 0:
-                    experiment.log_metric("z0_loss",result['split_latent_loss'],step=total_steps,epoch=epoch)
-                else:
-                    experiment.log_metric("z0_loss",result['position_final_loss'],step=total_steps,epoch=epoch)
-                experiment.log_metric("loss",result['loss'],step=total_steps,epoch=epoch)
-                experiment.log_metric("assoc_loss",result['association_final_loss'],step=total_steps,epoch=epoch)
-
-            elif trainable == "FullNetwork":
-                experiment.log_metric("loss",result['loss'],step=total_steps,epoch=epoch)
+            if nlatent > 0:
                 experiment.log_metric("z0_loss",result['split_latent_loss'],step=total_steps,epoch=epoch)
-                experiment.log_metric("assoc_loss",result['association_final_loss'],step=total_steps,epoch=epoch)
+            else:
+                experiment.log_metric("z0_loss",result['position_final_loss'],step=total_steps,epoch=epoch)
+            experiment.log_metric("loss",result['loss'],step=total_steps,epoch=epoch)
+            experiment.log_metric("assoc_loss",result['association_final_loss'],step=total_steps,epoch=epoch)
 
             if step%10==0:
-                predictedZ0_FH = eval.predictFastHisto(batch[z0],batch['trk_pt'])
-                if bit:
-                    predictedZ0_NN, predictedAssoc_NN,predicted_weights = model.predict_on_batch( [batch[bit_z0],WeightFeatures,trackFeatures] )
-                else:
-                    predictedZ0_NN, predictedAssoc_NN,predicted_weights = model.predict_on_batch( [batch[z0],WeightFeatures,trackFeatures] )
+                predictedZ0_FH = eval_funcs.predictFastHisto(batch[z0],batch['trk_pt'],return_index=bit)
+  
+                predictedZ0_NN, predictedAssoc_NN,predicted_weights = model.predict_on_batch( [batch[z0],WeightFeatures,trackFeatures] )
 
                 qz0_NN = np.percentile(predictedZ0_NN-batch['pvz0'],[5,32,50,68,95])
                 qz0_FH = np.percentile(predictedZ0_FH-batch['pvz0'],[5,32,50,68,95])
-
-                if trainable == "FH":
+         
+                if nlatent > 0:
                     print ("Step %02i-%02i: loss=%.3f (z0=%.3f, assoc=%.3f), q68=(%.4f,%.4f), [FH: q68=(%.4f,%.4f)]"%(
                             epoch,step,
-                            result['loss'],result['binto_vertex_loss'],result['association_final_loss'],
+                            result['loss'],result['split_latent_loss'],result['association_final_loss'],
                             qz0_NN[1],qz0_NN[3],qz0_FH[1],qz0_FH[3]
-                        ))
-                    print ("Train_NN_z0_MSE: "+str(metrics.mean_squared_error(batch['pvz0'],predictedZ0_NN))+" Train_FH_z0_MSE: "+str(metrics.mean_squared_error(batch['pvz0'],predictedZ0_FH)))          
-                elif ((trainable == "DiffArgMax") | (trainable == "QDiffArgMax")) :
-                    if nlatent > 0:
-                        print ("Step %02i-%02i: loss=%.3f (z0=%.3f, assoc=%.3f), q68=(%.4f,%.4f), [FH: q68=(%.4f,%.4f)]"%(
-                                epoch,step,
-                                result['loss'],result['split_latent_loss'],result['association_final_loss'],
-                                qz0_NN[1],qz0_NN[3],qz0_FH[1],qz0_FH[3]
-                            ))
-                    else:
-                        print ("Step %02i-%02i: loss=%.3f (z0=%.3f, assoc=%.3f), q68=(%.4f,%.4f), [FH: q68=(%.4f,%.4f)]"%(
-                                epoch,step,
-                                result['loss'],result['position_final_loss'],result['association_final_loss'],
-                                qz0_NN[1],qz0_NN[3],qz0_FH[1],qz0_FH[3]
-                            ))
-                    print ("Train_NN_z0_MSE: "+str(metrics.mean_squared_error(batch['pvz0'],predictedZ0_NN))+" Train_FH_z0_MSE: "+str(metrics.mean_squared_error(batch['pvz0'],predictedZ0_FH)))   
-                elif trainable == "FullNetwork":
-                    print ("Step %02i-%02i: loss=%.3f (split latent z0=%.3f, assoc=%.3f), q68=(%.4f,%.4f), [FH: q68=(%.4f,%.4f)]"%(
-                        epoch,step,
-                        result['loss'],result['split_latent_loss'],result['association_final_loss'],
-                        qz0_NN[1],qz0_NN[3],qz0_FH[1],qz0_FH[3]
                     ))
-
+                else:
+                    print ("Step %02i-%02i: loss=%.3f (z0=%.3f, assoc=%.3f), q68=(%.4f,%.4f), [FH: q68=(%.4f,%.4f)]"%(
+                            epoch,step,
+                            result['loss'],result['position_final_loss'],result['association_final_loss'],
+                            qz0_NN[1],qz0_NN[3],qz0_FH[1],qz0_FH[3]
+                    ))
+                print ("Train_NN_z0_MSE: "+str(metrics.mean_squared_error(batch['pvz0'],predictedZ0_NN))+" Train_FH_z0_MSE: "+str(metrics.mean_squared_error(batch['pvz0'],predictedZ0_FH)))   
 
             total_steps += 1
+
+        prune_level = []
+        for i,layer in enumerate(model.layers):
+            get_weights = layer.get_weights()
+            if len(get_weights) > 0:
+                if "Bin_weight" not in layer.name:
+                    weights = get_weights[0].flatten()[get_weights[0].flatten() != 0]
+                    prune_level.append(weights.shape[0]/get_weights[0].flatten().shape[0])
 
         val_actual_PV = []
         val_predictedZ0_FH = []
@@ -190,18 +161,13 @@ def train_model(model,experiment,train_files,val_files,trackfeat,weightfeat,epoc
         val_actual_assoc = []
 
         for val_step,val_batch in enumerate(setup_pipeline(val_files)):
-            val_predictedZ0_FH.append(eval.predictFastHisto(val_batch[z0],val_batch['trk_pt']).flatten())
+            val_predictedZ0_FH.append(eval_funcs.predictFastHisto(val_batch[z0],val_batch['trk_pt']).flatten())
 
             val_trackFeatures = np.stack([val_batch[feature] for feature in trackfeat],axis=2)
 
             val_WeightFeatures = np.stack([val_batch[feature] for feature in weightfeat],axis=2)
 
-            if bit:
-                temp_predictedZ0_NN, temp_predictedAssoc_NN,predicted_weights  = model.predict_on_batch(
-                        [val_batch[bit_z0],val_WeightFeatures,val_trackFeatures]
-                )
-            else:
-                temp_predictedZ0_NN, temp_predictedAssoc_NN,predicted_weights  = model.predict_on_batch(
+            temp_predictedZ0_NN, temp_predictedAssoc_NN,predicted_weights  = model.predict_on_batch(
                         [val_batch[z0],val_WeightFeatures,val_trackFeatures]
                 )
   
@@ -210,7 +176,7 @@ def train_model(model,experiment,train_files,val_files,trackfeat,weightfeat,epoc
             val_actual_PV.append(val_batch['pvz0'].numpy().flatten()) 
             val_actual_assoc.append(val_batch["trk_fromPV"].numpy().flatten())
 
-            val_predictedAssoc_FH.append(eval.FastHistoAssoc(val_batch['pvz0'],val_batch[z0],val_batch['trk_eta'],kf=kf).flatten())
+            val_predictedAssoc_FH.append(eval_funcs.FastHistoAssoc(val_batch['pvz0'],val_batch[z0],val_batch['trk_eta'],kf=kf).flatten())
         val_z0_NN_array = np.concatenate(val_predictedZ0_NN).ravel()
         val_z0_FH_array = np.concatenate(val_predictedZ0_FH).ravel()
         val_z0_PV_array = np.concatenate(val_actual_PV).ravel()
@@ -233,22 +199,23 @@ def train_model(model,experiment,train_files,val_files,trackfeat,weightfeat,epoc
         experiment.log_metric("Validation_FH_PV_ROC",metrics.roc_auc_score(val_assoc_PV_array,val_assoc_FH_array))
         experiment.log_metric("Validation_FH_PV_ACC",metrics.balanced_accuracy_score(val_assoc_PV_array,val_assoc_FH_array))
 
+        experiment.log_metric("Validation_Prune_Level",np.mean(prune_level))
+        
         old_lr = callbacks.on_epoch_end(epoch=epoch,logs=result,lr=new_lr)
 
         val_loss = metrics.mean_squared_error(val_z0_PV_array,val_z0_NN_array)
 
         print ("Val_NN_z0_MSE: "+str(val_loss)+" Best_NN_z0_MSE: "+str(best_score)) 
         
-
         wait += 1
         if val_loss < best_score:
             best_score = val_loss
-            model.save_weights(model_name+".tf")
+            model.save_weights(model_name[0]+model_name[1]+".tf")
             wait = 0
         if wait >= early_stop_patience:
             break
         
-def test_model(model,experiment,test_files,trackfeat,weightfeat,model_name=None):
+def test_model(model,experiment,test_files,trackfeat,weightfeat,model_name=[None,None]):
 
     predictedZ0_FH = []
     predictedZ0_NN = []
@@ -264,17 +231,14 @@ def test_model(model,experiment,test_files,trackfeat,weightfeat,model_name=None)
         trackFeatures = np.stack([batch[feature] for feature in trackfeat],axis=2)
         WeightFeatures = np.stack([batch[feature] for feature in weightfeat],axis=2)
 
-        predictedZ0_FH.append(eval.predictFastHisto(batch[z0],batch['trk_pt']))
+        predictedZ0_FH.append(eval_funcs.predictFastHisto(batch[z0],batch['trk_pt']))
 
         actual_Assoc.append(batch["trk_fromPV"])
         actual_PV.append(batch['pvz0'])
 
-        predictedAssoc_FH.append(eval.FastHistoAssoc(batch['pvz0'],batch[z0],batch['trk_eta'],kf=kf))
+        predictedAssoc_FH.append(eval_funcs.FastHistoAssoc(batch['pvz0'],batch[z0],batch['trk_eta'],kf=kf))
 
-        if bit:   
-            predictedZ0_NN_temp, predictedAssoc_NN_temp,predicted_weights = model.predict_on_batch( [batch[bit_z0],WeightFeatures,trackFeatures] )
-        else:
-            predictedZ0_NN_temp, predictedAssoc_NN_temp,predicted_weights = model.predict_on_batch( [batch[z0],WeightFeatures,trackFeatures] )
+        predictedZ0_NN_temp, predictedAssoc_NN_temp,predicted_weights = model.predict_on_batch( [batch[z0],WeightFeatures,trackFeatures] )
 
         predictedZ0_NN.append(predictedZ0_NN_temp)
         predictedAssoc_NN.append(predictedAssoc_NN_temp)
@@ -290,8 +254,8 @@ def test_model(model,experiment,test_files,trackfeat,weightfeat,model_name=None)
     qz0_NN = np.percentile(z0_NN_array-z0_PV_array,[5,15,50,85,95])
     qz0_FH = np.percentile(z0_FH_array-z0_PV_array,[5,15,50,85,95])
 
-    experiment.log_asset(model_name+".tf.index")
-    experiment.log_asset(model_name+".tf.data-00000-of-00001")
+    experiment.log_asset(model_name[0]+model_name[1]+".tf.index")
+    experiment.log_asset(model_name[0]+model_name[1]+".tf.data-00000-of-00001")
 
     experiment.log_metric("Test_NN_z0_MSE",metrics.mean_squared_error(z0_PV_array,z0_NN_array))
     experiment.log_metric("Test_NN_z0_AE",metrics.mean_absolute_error(z0_PV_array,z0_NN_array))
@@ -322,21 +286,30 @@ if __name__=="__main__":
         config = yaml.load(f,Loader=yaml.FullLoader)
     retrain = config["retrain"]
 
-    trainable = config["trainable"]
+    trainable = sys.argv[3]
     trackfeat = config["track_features"] 
     weightfeat = config["weight_features"] 
-    bit = config["bit_inputs"]
+    
     nlatent = config["Nlatent"]
     PretrainedModelName = config["PretrainedModelName"] 
-    UnQuantisedModelName = config["UnquantisedModelName"] 
     pretrain_DA = config["pretrain_DA"]
 
-    if (trainable == "DiffArgMax") | (trainable == "QDiffArgMax"):
-        
+    if (kf == 'NewKF') | (kf == 'OldKF'):
+        start = -15
+        end = 15
+        bit = False
+
+    elif (kf == 'OldKF_intZ') :
+        start = 0
+        end = 256
+        bit = True
+
+    if trainable == 'DA':
+
         network = vtx.nn.E2EDiffArgMax(
             nbins=256,
-            start=-15,
-            end = 15,
+            start=start,
+            end = end,
             ntracks=max_ntracks, 
             nweightfeatures=len(weightfeat), 
             nfeatures=len(trackfeat), 
@@ -351,54 +324,68 @@ if __name__=="__main__":
             nassoclayers = config['nassoclayers'],
         )
 
-    elif trainable == "FH":
+        model_name = [config["UnquantisedModelName"] ,""]
 
-        network = vtx.nn.E2EFH(
+    if trainable == 'QDA':
+
+        network = vtx.nn.E2EQKerasDiffArgMax(
             nbins=256,
-            ntracks=max_ntracks, 
-            nweightfeatures=len(weightfeat), 
-            nfeatures=len(trackfeat), 
-            nweights=1, 
-            activation='relu',
-            regloss=1e-10
-        )
-
-    elif trainable == "FullNetwork":
-
-        network = vtx.nn.E2Ecomparekernel(
-            nbins=256,
+            start = start,
+            end = end,
             ntracks=max_ntracks, 
             nweightfeatures=len(weightfeat), 
             nfeatures=len(trackfeat), 
             nweights=1, 
             nlatent = nlatent,
             activation='relu',
-            regloss=1e-10
+            l1regloss = (float)(config['l1regloss']),
+            l2regloss = (float)(config['l2regloss']),
+            nweightnodes = config['nweightnodes'],
+            nweightlayers = config['nweightlayers'],
+            nassocnodes = config['nassocnodes'],
+            nassoclayers = config['nassoclayers'],
+            temperature = 1e-2,
+            qconfig = config['QConfig']
         )
+
+        model_name = [config['QuantisedModelName'],"_prune_iteration_0"]
+
+    if trainable == 'QDA_prune':
+        network = vtx.nn.E2EQKerasDiffArgMaxConstraint(
+            nbins=256,
+            start = start,
+            end = end,
+            ntracks=max_ntracks, 
+            nweightfeatures=len(weightfeat), 
+            nfeatures=len(trackfeat), 
+            nweights=1, 
+            nlatent = nlatent,
+            activation='relu',
+            l1regloss = (float)(config['l1regloss']),
+            l2regloss = (float)(config['l2regloss']),
+            nweightnodes = config['nweightnodes'],
+            nweightlayers = config['nweightlayers'],
+            nassocnodes = config['nassocnodes'],
+            nassoclayers = config['nassoclayers'],
+            temperature = 1e-2,
+            qconfig = config['QConfig'],
+            h5fName = config['QuantisedModelName']+'_drop_weights_iteration_'+sys.argv[4]+'.h5'
+        )
+
+        model_name = [config['QuantisedModelName'],"_prune_iteration_"+str(int(sys.argv[3])+1)]
+    
 
     if kf == "NewKF":
         train_files = glob.glob(config["data_folder"]+"/Train/*.tfrecord")
         test_files = glob.glob(config["data_folder"]+"/Test/*.tfrecord")
         val_files = glob.glob(config["data_folder"]+"/Val/*.tfrecord")
-
-        trackFeatures = [
-            'trk_z0',
-            'normed_trk_pt',
-            'normed_trk_eta',
-            'trk_pt',
-            'trk_eta',
-            'trk_MVA1',
-            'trk_z0_res',
-            'corrected_trk_z0',
-            'normed_trk_over_eta',
-    ]
         
-    elif kf == "OldKF":
+    elif (kf == "OldKF") | (kf == 'OldKF_intZ'):
         train_files = glob.glob(config["data_folder"]+"/Train/*.tfrecord")
         test_files = glob.glob(config["data_folder"]+"/Test/*.tfrecord")
         val_files = glob.glob(config["data_folder"]+"/Val/*.tfrecord")
 
-        trackFeatures = [
+    trackFeatures = [
             'trk_z0',
             'normed_trk_pt',
             'normed_trk_eta',
@@ -407,19 +394,20 @@ if __name__=="__main__":
             'trk_MVA1',
             'trk_z0_res',
             'corrected_trk_z0',
+            'corrected_int_z0',
             'normed_trk_over_eta',
             'abs_trk_word_pT',
             'rescaled_trk_word_MVAquality',
             'abs_trk_word_eta',
     ]
-       
+
 
     print ("Input Train files: ",len(train_files))
     print ("Input Validation files: ",len(val_files))
     print ("Input Test files: ",len(test_files))
 
     features = {
-        "pvz0": tf.io.FixedLenFeature([1], tf.float32),    
+        "pvz0": tf.io.FixedLenFeature([1], tf.float32),  
         "trk_fromPV":tf.io.FixedLenFeature([max_ntracks], tf.float32) ,
         "PV_hist"  :tf.io.FixedLenFeature([256,1], tf.float32),
     }
@@ -444,22 +432,13 @@ if __name__=="__main__":
     startingLR = config['starting_lr']
     epochs = config['epochs']
 
-    experiment.log_parameter("nbins",256)
-    experiment.log_parameter("ntracks",max_ntracks)
-    experiment.log_parameter("nfeatures",3)
-    experiment.log_parameter("nlatent",2)
-    experiment.log_parameter("activation",'relu')
-    experiment.log_parameter("regloss",1e-10)
-    experiment.log_parameter("Start LR",startingLR)
-    experiment.log_parameter("Epochs",epochs)
-
     model = network.createE2EModel()
     optimizer = tf.keras.optimizers.Adam(lr=startingLR)
     model.compile(
         optimizer,
         loss=[
-            #tf.keras.losses.Huber(config['Huber_delta']),
-            tf.keras.losses.MeanAbsoluteError(),
+            tf.keras.losses.Huber(config['Huber_delta']),
+            #tf.keras.losses.MeanAbsoluteError(),
             tf.keras.losses.BinaryCrossentropy(from_logits=True),
             lambda y,x: 0.,
         ],
@@ -473,27 +452,70 @@ if __name__=="__main__":
     )
     model.summary()
 
-    if trainable == "FH":
+    model.layers[11].set_weights([np.expand_dims(np.arange(256),axis=0)]) #Set to bin index 
 
-        model.layers[2].set_weights([np.array([[1]], dtype=np.float32)])
-        model.layers[4].set_weights([np.array([[[1]],[[1]],[[1]]], dtype=np.float32)])
-        model.layers[7].set_weights([np.expand_dims(np.arange(256),axis=0)])
-    elif trainable == "DiffArgMax":
-        model.layers[17].set_weights([np.expand_dims(np.arange(256),axis=0)]) #Set to bin index 
-        model.layers[19].set_weights([np.array([[1]], dtype=np.float32), np.array([0], dtype=np.float32)])
+    if trainable == 'DA':
+        if pretrain_DA:
+            model.load_weights(PretrainedModelName+".tf")
+    elif trainable == 'QDA':
+        if config["pretrained"]:
+            DAnetwork = vtx.nn.E2EDiffArgMax(
+                nbins=256,
+                ntracks=max_ntracks, 
+                nweightfeatures=len(weightfeat), 
+                nfeatures=len(trackfeat), 
+                nweights=1, 
+                nlatent = nlatent,
+                activation='relu',
+                l2regloss=1e-10,
+                nweightnodes = config['nweightnodes'],
+                nweightlayers = config['nweightlayers'],
+                nassocnodes = config['nassocnodes'],
+                nassoclayers = config['nassoclayers'],
+            )
+            
+            DAmodel = DAnetwork.createE2EModel()
+            optimizer = tf.keras.optimizers.Adam(lr=0.01)
+            DAmodel.compile(
+                optimizer,
+                loss=[
+                    tf.keras.losses.Huber(config['Huber_delta']),
+                    #tf.keras.losses.MeanAbsoluteError(),
+                    tf.keras.losses.BinaryCrossentropy(from_logits=True),
+                    lambda y,x: 0.
+                ],
+                metrics=[
+                    tf.keras.metrics.BinaryAccuracy(threshold=0.,name='assoc_acc') #use thres=0 here since logits are used
+                ],
+                loss_weights=[config['z0_loss_weight'],
+                            config['crossentropy_loss_weight'],
+                            0]
+            )
+            DAmodel.summary()
+            DAmodel.load_weights(config["UnquantisedModelName"]+".tf")
 
-    elif trainable == "QDiffArgMax":
-        model.layers[11].set_weights([np.expand_dims(np.arange(256),axis=0)]) #Set to bin index 
+            model.layers[1].set_weights(DAmodel.layers[1].get_weights())
+            model.layers[3].set_weights(DAmodel.layers[3].get_weights()) 
+            model.layers[5].set_weights(DAmodel.layers[6].get_weights()) 
+            model.layers[9].set_weights(DAmodel.layers[8].get_weights()) 
+            model.layers[13].set_weights(DAmodel.layers[11].get_weights()) 
+            model.layers[15].set_weights(DAmodel.layers[13].get_weights()) 
+            model.layers[19].set_weights(DAmodel.layers[17].get_weights()) 
+            model.layers[21].set_weights(DAmodel.layers[19].get_weights()) 
+            model.layers[23].set_weights(DAmodel.layers[21].get_weights())
 
-    if pretrain_DA:
-        model.load_weights(PretrainedModelName+".tf")
-        
+        else:
+            model.layers[13].set_weights([np.expand_dims(np.arange(256),axis=0)]) #Set to bin index 
+
+    elif trainable == 'QDA_prune':
+        model.load_weights(config["QuantisedModelName"]+"_prune_iteration_"+sys.argv[4]+".tf").expect_partial()
+
 
     reduceLR = TrainingScripts.Callbacks.OwnReduceLROnPlateau(
-    monitor='loss', factor=0.1, patience=6, verbose=1,
-    mode='auto', min_delta=0.005, cooldown=0, min_lr=0
+        monitor='loss', factor=0.1, patience=6, verbose=1,
+        mode='auto', min_delta=0.005, cooldown=0, min_lr=0
     )
     with experiment.train():
-        train_model(model,experiment,train_files,val_files,trackfeat,weightfeat,epochs=epochs,callbacks=reduceLR,nlatent=nlatent,trainable=trainable,bit=bit,model_name=UnQuantisedModelName),
+        train_model(model,experiment,train_files,val_files,trackfeat,weightfeat,epochs=epochs,callbacks=reduceLR,nlatent=nlatent,bit=bit,model_name=model_name),
     with experiment.test():
-        test_model(model,experiment,test_files,trackfeat,weightfeat,model_name=UnQuantisedModelName)
+        test_model(model,experiment,test_files,trackfeat,weightfeat,model_name=model_name,bit=bit)
