+++ Train/vtx/nn/E2EQKerasDiffArgMax_Constraint.py
@@ -279,8 +279,6 @@ class E2EQKerasDiffArgMaxConstraint():
         cmsml.tensorflow.save_graph(modelName+"_patternModelgraph.pb", self.patternModel, variables_to_constants=True)
         cmsml.tensorflow.save_graph(modelName+"_associationModelgraph.pb", self.associationModel, variables_to_constants=True)
 
-        
-
     def export_individual_models(self,modelName):
 
         with open(modelName+"_weightModel.json", 'w') as f:
@@ -315,16 +313,16 @@ class E2EQKerasDiffArgMaxConstraint():
         hls_weight_model = hls4ml.converters.convert_from_keras_model(self.weightModel,
                                                             hls_config=weightconfig,
                                                             output_dir='/home/cebrown/Documents/Trigger/E2E/Train/'+modelName+'_hls_weight/hls4ml_prj',
-                                                            fpga_part='xcvu9p-flga2104-2L-e',
-                                                            clock_period=2.5)
-        hls4ml.utils.plot_model(hls_weight_model, show_shapes=True, show_precision=True, to_file=modelName+"_Weight_model.png")
-        plt.clf()
-        ap, wp = hls4ml.model.profiling.numerical(model=self.weightModel, hls_model=hls_weight_model, X=random_weight_data)
-        wp.savefig(modelName+"_Weight_model_activations_profile.png")
-        ap.savefig(modelName+"_Weight_model_weights_profile.png")
+                                                            part='xcvu9p-flga2104-2L-e',
+                                                            clock_period=1.8)
+        #hls4ml.utils.plot_model(hls_weight_model, show_shapes=True, show_precision=True, to_file=modelName+"_Weight_model.png")
+        #plt.clf()
+        #ap, wp = hls4ml.model.profiling.numerical(model=self.weightModel, hls_model=hls_weight_model, X=random_weight_data)
+        #wp.savefig(modelName+"_Weight_model_activations_profile.png")
+        #ap.savefig(modelName+"_Weight_model_weights_profile.png")
 
         hls_weight_model.compile()
-        hls_weight_model.build(csim=False,synth=True,vsynth=True)
+        hls_weight_model.build(csim=True,synth=True,vsynth=True)
 
 
 
@@ -346,7 +344,7 @@ class E2EQKerasDiffArgMaxConstraint():
                                                             hls_config=patternconfig,
                                                             output_dir='/home/cebrown/Documents/Trigger/E2E/Train/'+modelName+'_hls_pattern/hls4ml_prj',
                                                             fpga_part='xcvu9p-flga2104-2L-e',
-                                                            clock_period=2.5)
+                                                            clock_period=1.8)
         #hls4ml.utils.plot_model(hls_pattern_model, show_shapes=True, show_precision=True, to_file=modelName+"_pattern_model.png")
         #plt.clf()
         #ap,wp = hls4ml.model.profiling.numerical(model=self.patternModel, hls_model=hls_pattern_model, X=random_pattern_data)
@@ -354,7 +352,7 @@ class E2EQKerasDiffArgMaxConstraint():
         #ap.savefig(modelName+"_pattern_model_weights_profile.png")
 
         hls_pattern_model.compile()
-        hls_pattern_model.build(csim=False,synth=True,vsynth=True)
+        hls_pattern_model.build(csim=True,synth=True,vsynth=True)
 
     def export_hls_assoc_model(self,modelName):
         import hls4ml
@@ -373,13 +371,13 @@ class E2EQKerasDiffArgMaxConstraint():
         hls_association_model = hls4ml.converters.convert_from_keras_model(self.associationModel,
                                                             hls_config=associationconfig,
                                                             output_dir='/home/cebrown/Documents/Trigger/E2E/Train/'+modelName+'_hls_association/hls4ml_prj',
-                                                            fpga_part='xcvu9p-flga2104-2L-e',
-                                                            clock_period=2.5)
-        hls4ml.utils.plot_model(hls_association_model, show_shapes=True, show_precision=True, to_file=modelName+"_association_model.png")
-        plt.clf()
-        ap,wp = hls4ml.model.profiling.numerical(model=self.associationModel, hls_model=hls_association_model, X=random_association_data)
-        wp.savefig(modelName+"_association_model_activations_profile.png")
-        ap.savefig(modelName+"_association_model_weights_profile.png")
+                                                            part='xcvu9p-flga2104-2L-e',
+                                                            clock_period=1.8)
+        #hls4ml.utils.plot_model(hls_association_model, show_shapes=True, show_precision=True, to_file=modelName+"_association_model.png")
+        #plt.clf()
+        #ap,wp = hls4ml.model.profiling.numerical(model=self.associationModel, hls_model=hls_association_model, X=random_association_data)
+        #wp.savefig(modelName+"_association_model_activations_profile.png")
+        #ap.savefig(modelName+"_association_model_weights_profile.png")
 
         hls_association_model.compile()
-        hls_association_model.build(csim=False,synth=True,vsynth=True)
+        hls_association_model.build(csim=True,synth=True,vsynth=True)
